# src/backend/routes/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Drivers Logbook Web Application API",
    "version": "1.0.0",
    "description": "API documentation for the Drivers Logbook Web Application."
  },
  "servers": [
    {
      "url": "http://localhost/",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/report": {
      "get": {
        "summary": "Serves the report page.",
        "description": "Returns the report HTML page. Authentication is required.",
        "tags": [
          "Static Pages"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "HTML file for the report page."
          }
        }
      }
    },
...
    "/api/trips": {
      "get": {
        "summary": "Retrieves trips.",
        "description": "Returns a list of trips. Admins can view all trips. managers can only view their own vehicle's trips. dispatchers can view all business trips.",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "vehicleId",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter for vehicle ID."
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of trips."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "summary": "Deletes multiple trips.",
        "description": "Deletes (or marks as deleted) multiple trips by trip IDs. Requires admin API key for full deletion.",
        "tags": [
          "Trips"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "Object with an array of trip IDs.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tripIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trips deleted or marked as deleted."
          },
          "400": {
            "description": "No trip IDs provided."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/trip/{id}": {
      "get": {
        "summary": "Retrieves a specific trip.",
        "description": "Returns a trip identified by its ID.",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Trip ID."
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A single trip."
          },
          "404": {
            "description": "Trip not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "patch": {
        "summary": "Updates a trip.",
        "description": "Updates an existing trip. Editing is allowed only if the trip is within the permitted time span.",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Trip ID."
          }
        ],
        "requestBody": {
          "description": "Trip data to update.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tripNotes": {
                    "type": "string"
                  },
                  "clientCompany": {
                    "type": "string"
                  },
                  "client": {
                    "type": "string"
                  },
                  "detourNote": {
                    "type": "string"
                  },
                  "tripCategory": {
                    "type": "string"
                  },
                  "tripPurpose": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trip updated successfully."
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Editing not allowed."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "summary": "Deletes a trip.",
        "description": "Deletes (or marks as deleted) a specific trip. Admins delete permanently.",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Trip ID."
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Trip deleted or marked as deleted."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/trip": {
      "post": {
        "summary": "Creates a new trip.",
        "description": "Creates a new trip entry.",
        "tags": [
          "Trips"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startLocation": {
                    "type": "string"
                  },
                  "endLocation": {
                    "type": "string"
                  },
                  "startTimestamp": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endTimestamp": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "startMileage": {
                    "type": "number"
                  },
                  "endMileage": {
                    "type": "number"
                  },
                  "tripCategory": {
                    "type": "string"
                  },
                  "tripPurpose": {
                    "type": "string"
                  },
                  "tripNotes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "tripStatus": {
                    "type": "string"
                  },
                  "recorded": {
                    "type": "boolean"
                  },
                  "vehicle": {
                    "type": "object",
                    "properties": {
                      "vin": {
                        "type": "string"
                      },
                      "manufacturer": {
                        "type": "string"
                      },
                      "year": {
                        "type": "number"
                      },
                      "region": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trip created successfully."
          },
          "400": {
            "description": "Bad request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/trips/merge": {
      "post": {
        "summary": "Merges multiple trips.",
        "description": "Merges the trips specified by an array of trip IDs. All trips must be editable and valid.",
        "tags": [
          "Trips"
        ],
        "requestBody": {
          "description": "Object containing an array of trip IDs.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tripIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trips merged successfully."
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Forbidden merge."
          }
        }
      }
    },
    "/api/report": {
      "get": {
        "summary": "Generates a trip report.",
        "description": "Generates a report based on a date range provided as query parameters.",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "fromDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date."
          },
          {
            "in": "query",
            "name": "toDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date."
          }
        ],
        "responses": {
          "200": {
            "description": "Report data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array"
                }
              }
            }
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
...
}